upgrade: move index page to App router and style
