+++ app/page.js
@@ -1,0 +1,15 @@
+import Activity from "../components/activity";
+import Hero from "../components/hero";
+import Partner from "../components/partner";
+import Information from "../components/information";
+
+export default function Home() {
+  return (
+    <>
+      <Hero />
+      <Activity />
+      <Partner />
+      <Information />
+    </>
+  );
+}
