+++ components/navbar.js
@@ -1,3 +1,4 @@
+"use client";
 import React, { useState } from "react";
 
 const joinUsLink = "https://forms.gle/L3oGiB5EwVTTJmZ9A";
@@ -7,7 +8,7 @@ export default function Navbar() {
 
   return (
     <>
-      <nav className="w-full z-10 flex items-center justify-between p-6 bg-white">
+      <nav className="w-full z-10 flex items-center justify-between p-6 bg-white container mx-auto">
         {/* Logo */}
         <div className="flex justify-center w-full md:w-auto">
           <a href="/">
@@ -16,7 +17,7 @@ export default function Navbar() {
         </div>
 
         {/* Navigation menu for larger screens */}
-        <div className="hidden md:flex space-x-12 text-4xl font-semibold">
+        <div className="hidden md:flex space-x-8 text-2xl font-semibold">
           <a href="/" className="text-red-400 hover:text-red-600">
             Home
           </a>
