+++ package.json
@@ -12,7 +12,7 @@
     "date-fns": "^2.29.3",
     "gray-matter": "^4.0.3",
     "markdown-it": "^13.0.1",
-    "next": "^13.4.7",
+    "next": "^14.2.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-loading": "^2.0.3",
@@ -24,8 +24,10 @@
   },
   "devDependencies": {
     "@tailwindcss/typography": "^0.5.9",
+    "@types/react": "18.3.3",
     "autoprefixer": "^10.4.13",
     "postcss": "^8.4.21",
-    "tailwindcss": "^3.2.4"
+    "tailwindcss": "^3.2.4",
+    "typescript": "5.4.5"
   }
 }
